Changed the behavior when passing array to <select>
If you pass an array to a <select> (not <select multiple>), it
uses the first element of the array for selection criteria, rather
than the last.

Initially it was set to the last index because I thought that'd
kind of represent the final option selected.  But just using index
0 is cleaner, and you shouldn't pass an array to a non multiple
select element anyway.